Backend: Reorganized Saga Requests and commands
