+++ FinanceBackEnd/Finance.Application/Services/Interfaces/ISagaService.cs
@@ -1,5 +1,6 @@
 using Finance.Domain.Models.Interfaces;
 using Microsoft.EntityFrameworkCore.Storage;
+using ExecutionContext = CQRSDispatch.ExecutionContext;
 
 namespace Finance.Application.Services.Interfaces;
 
@@ -9,7 +10,7 @@ public interface ISagaService<TCreateRequest, TUpdateRequest, TDeleteRequest, TE
     where TDeleteRequest : ISagaRequest
     where TEntity : IEntity?
 {
-    Task<(TEntity result, bool success, string error)> Create(TCreateRequest request, IDbContextTransaction? transaction = null);
-    Task<(TEntity result, bool success, string error)> Update(TUpdateRequest request, IDbContextTransaction? transaction = null);
-    Task<(bool success, string error)> Delete(TDeleteRequest request, IDbContextTransaction? transaction = null);
+    Task<(TEntity result, bool success, string error)> Create(TCreateRequest request, IDbContextTransaction? transaction = null, ExecutionContext? executionContext = null);
+    Task<(TEntity result, bool success, string error)> Update(TUpdateRequest request, IDbContextTransaction? transaction = null, ExecutionContext? executionContext = null);
+    Task<(bool success, string error)> Delete(TDeleteRequest request, IDbContextTransaction? transaction = null, ExecutionContext? executionContext = null);
 }
