+++ FinanceBackEnd/Finance.Application/Services/Requests/FundOwners/BaseFundOwnerSagaRequest.cs
@@ -2,7 +2,7 @@ using CQRSDispatch;
 using Finance.Application.Commands.Users;
 using Finance.Application.Services.Interfaces;
 
-namespace Finance.Application.Commands;
+namespace Finance.Application.Commands.FundOwners;
 
 public abstract class BaseFundOwnerSagaRequest<TResult> : OwnerBaseCommand<TResult>, ISagaRequest
     where TResult : RequestResult
