+++ FinanceBackEnd/Finance.Application/Services/Requests/FundOwners/SetFundOwnerSagaRequest.cs
@@ -2,7 +2,7 @@ using CQRSDispatch;
 using Finance.Application.Services.Interfaces;
 using Finance.Domain.Models;
 
-namespace Finance.Application.Commands;
+namespace Finance.Application.Commands.FundOwners;
 
 public class SetFundOwnerSagaRequest : BaseFundOwnerSagaRequest<DataResult<FundResource>>, ISagaRequest
 {
